+++ controller/src/main/java/com/platon/metis/admin/dto/resp/LocalDataDetailResp.java
@@ -56,7 +56,7 @@ public class LocalDataDetailResp {
     private Long size;
     //数据行数(不算title)
     @ApiModelProperty(name = "rows", value = "数据行数(不算title)")
-    private Long rows;
+    private Integer rows;
     //数据列数
     @ApiModelProperty(name = "columns", value = "数据列数")
     private Integer columns;
@@ -102,6 +102,7 @@ public class LocalDataDetailResp {
         resp.setRemarks(localMetaData.getRemarks());
         resp.setIndustry(localMetaData.getIndustry());
         resp.setMetaDataId(localMetaData.getMetaDataId());
+
         resp.setAttendTaskCount((Integer)localMetaData.getField("taskCount"));
         resp.setResourceName(localMetaData.getMetaDataName());
         resp.setId(localMetaData.getId());
